renamed variables subfolder to parameterization
